+++ docs/changelog.md
@@ -14,6 +14,10 @@ Each revision is versioned by the date of the revision.
 
 - Added GitHub workflow that checks whether a pull request contains a change artifact.
 
+## 2025-11-12
+
+- Updated the haproxy-route library to add the `allow_http` attribute.
+
 ## 2025-10-14
 
 - Added action `get-proxied-endpoints`.